refactor: impl Index for Piece, Side

+++ src/board/make_move.rs
@@ -94,9 +94,7 @@ impl BoardState {
         self.en_passant_square = if m.move_type.is_double_push() {
             let t = m.target as usize;
             let adjacent = ((1u64 << (t - 1)) & !FILE_H) | ((1u64 << (t + 1)) & !FILE_A);
-            if (self.pieces[self.side_to_move.other() as usize][Piece::Pawn as usize] & adjacent)
-                .is_not_empty()
-            {
+            if (self.pieces[self.side_to_move.other()][Piece::Pawn] & adjacent).is_not_empty() {
                 self.en_passant_square_for(m)
             } else {
                 Square::NoSquare
@@ -189,15 +187,15 @@ impl BoardState {
     }
 
     pub fn is_draw(&self) -> bool {
-        let num_pieces = self.occupancies[Side::Both as usize].count_ones();
+        let num_pieces = self.occupancies[Side::Both].count_ones();
         if num_pieces == 2 {
             // Assumes a legal board with 2 Kings only
             return true;
         } else if num_pieces == 3 {
-            let knights = self.pieces[Side::White as usize][Piece::Knight as usize]
-                | self.pieces[Side::Black as usize][Piece::Knight as usize];
-            let bishops = self.pieces[Side::White as usize][Piece::Bishop as usize]
-                | self.pieces[Side::Black as usize][Piece::Bishop as usize];
+            let knights =
+                self.pieces[Side::White][Piece::Knight] | self.pieces[Side::Black][Piece::Knight];
+            let bishops =
+                self.pieces[Side::White][Piece::Bishop] | self.pieces[Side::Black][Piece::Bishop];
             if (knights | bishops).is_not_empty() {
                 return true;
             }
